Data source type definition. Adds function NewDataSource.
